# hackathonApp/backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon App API
  description: |
    REST API for the Hackathon Application. Supports user authentication (with optional
    MFA), ticket management, real-time chat (channels, DMs, ticket threads), Workday
    ad-hoc sync for admin users, and MFA administration.

    ## Authentication
    All authenticated endpoints require a valid session cookie (`hackathon_session`).
    Obtain a session by calling `POST /api/auth/login`. If the user has MFA enabled,
    a two-step login flow is required (see MFA section).

    ## Rate Limiting
    All endpoints are rate-limited. Unauthenticated endpoints are limited per IP;
    authenticated endpoints are limited per user. When rate-limited, the API returns
    `429 Too Many Requests` with a `Retry-After` header.

    ## Error Format
    All errors follow a consistent JSON format:
    ```json
    {
      "error": "error_code_or_message",
      "message": "Human-readable description"
    }
    ```
  version: 1.0.0
  contact:
    name: Hackathon App Team
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://hackathon.example.com/api
    description: Production server
...
paths:
...
  /chats/channels:
    get:
      tags: [Chat]
      summary: List all channels
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Channel list
          content:
            application/json:
              schema:
                type: object
                properties:
                  channels:
                    type: array
                    items:
                      $ref: '#/components/schemas/Channel'
    post:
      tags: [Chat]
      summary: Create a new channel (admin only)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '403':
          description: Admin role required
  /chats/channels/{channelId}/join:
    post:
      tags: [Chat]
      summary: Join a channel
      security:
        - sessionCookie: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Joined channel
  /chats/direct-messages:
    get:
      tags: [Chat]
      summary: List current user's DM conversations
      security:
        - sessionCookie: []
      responses:
        '200':
          description: DM conversation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/DMConversation'
    post:
      tags: [Chat]
      summary: Start or get existing DM conversation
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipient_id]
              properties:
                recipient_id:
                  type: integer
      responses:
        '200':
          description: DM conversation
  /chats/{conversationId}/messages:
    get:
      tags: [Chat]
      summary: Get paginated messages for a conversation
      security:
        - sessionCookie: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: before
          in: query
          description: Message ID cursor for pagination (load older messages)
          schema:
            type: integer
        - name: limit
          in: query
          description: Messages per page (max 100)
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Message list
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  has_more:
                    type: boolean
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hackathon_session
  schemas:
...
    Channel:
      type: object
      properties:
        id:
          type: string
          example: channel_general
        name:
          type: string
        description:
          type: string
        member_count:
          type: integer
        is_member:
          type: boolean
        unread_count:
          type: integer
        last_message:
          $ref: '#/components/schemas/MessagePreview'
    DMConversation:
      type: object
      properties:
        id:
          type: string
          example: dm_3_7
        type:
          type: string
          enum: [direct]
        other_user:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
        unread_count:
          type: integer
        last_message:
          $ref: '#/components/schemas/MessagePreview'
    Message:
      type: object
      properties:
        id:
          type: integer
        sender:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
        content:
          type: string
        conversation_id:
          type: string
        mentions:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        timestamp:
          type: string
          format: date-time
    MessagePreview:
      type: object
      properties:
        content:
          type: string
        sender_name:
          type: string
        timestamp:
          type: string
          format: date-time
...